requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
azure-cosmos
pydantic
pydantic-settings
gremlinpython==3.7.2
async-timeout
asyncio